Convert BattlePlanUpdate::update
