+++ Code/GameEngine/Source/GameLogic/Object/Update/BattlePlanUpdate.cpp
@@ -327,93 +327,6 @@ CommandOption BattlePlanUpdate::getCommandOption() const
 	return (CommandOption)0;
 }
 
-//-------------------------------------------------------------------------------------------------
-/** The update callback. */
-//-------------------------------------------------------------------------------------------------
-// ?update@BattlePlanUpdate@@UAE?AW4UpdateSleepTime@@XZ present-unmatched
-UpdateSleepTime BattlePlanUpdate::update()
-{	
-
-	if( m_invalidSettings )
-	{
-		// can't return UPDATE_SLEEP_FOREVER unless we are sleepy...
-		return UPDATE_SLEEP_NONE;
-		///return UPDATE_SLEEP_FOREVER;
-	}
-
-	//const BattlePlanUpdateModuleData *data = getBattlePlanUpdateModuleData();
-	//Object *obj = getObject();
-
-	UnsignedInt now = TheGameLogic->getFrame();
-
-	if( m_nextReadyFrame <= now )
-	{
-		switch( m_status )
-		{
-			case TRANSITIONSTATUS_IDLE:
-				//There's only two cases where we are in an idle status -- upon initialization
-				//when no plan has yet been selected. The other case is after we've finished
-				//packing the previous plan up and are waiting to unpack the new state.
-				if( m_desiredPlan != PLANSTATUS_NONE )
-				{
-					m_currentPlan = m_desiredPlan;
-					setStatus( TRANSITIONSTATUS_UNPACKING );
-				}
-				break;
-			case TRANSITIONSTATUS_UNPACKING:
-				//If we're unpacking, we are forcing the user to wait until the plan is unpacked
-				//before allowing him to select a new plan. The plan doesn't become active until
-				//we're finished unpacking.
-				setStatus( TRANSITIONSTATUS_ACTIVE );
-				if( m_currentPlan == PLANSTATUS_BOMBARDMENT )
-				{
-					enableTurret( true );
-				}
-				break;
-			case TRANSITIONSTATUS_ACTIVE:
-				//If we're active and the user has selected a different plan, then we need to 
-				//pack up.
-				if( m_currentPlan != m_desiredPlan )
-				{
-					if( m_currentPlan == PLANSTATUS_BOMBARDMENT )
-					{
-						//Special case situation -- in bombardment status, we need to center
-						//the turret prior to packing up, so handle it here.
-						AIUpdateInterface *ai = getObject()->getAI();
-						if( ai )
-						{
-							if( isTurretInNaturalPosition() )
-							{
-								//It's centered, so pack
-								setStatus( TRANSITIONSTATUS_PACKING );
-								m_centeringTurret = false;
-								enableTurret( false );
-							}
-							else if( !m_centeringTurret )
-							{
-								//It's not centered, and not trying to center, so order it to center.
-								ai->aiIdle( CMD_FROM_AI );
-								recenterTurret();
-								m_centeringTurret = true;
-							}
-						}
-					}
-					else
-					{
-						setStatus( TRANSITIONSTATUS_PACKING );
-					}
-				}
-				break;
-			case TRANSITIONSTATUS_PACKING:
-				//If we finished packing, then go idle until we can switch to our new plan.
-				setStatus( TRANSITIONSTATUS_IDLE );
-				break;
-		}
-	}
-
-	return UPDATE_SLEEP_NONE;
-}
-
 // ------------------------------------------------------------------------------------------------
 /** Create vision objects for all players revealing this building to all */
 // ------------------------------------------------------------------------------------------------
